Uniform admin remove command

+++ src/discordBot/commands/admin/remove.js
@@ -2,28 +2,30 @@ const { updateGuide, findCategoryName, removeCourseFromDb } = require("../../ser
 const { courseAdminRole } = require("../../../../config.json");
 
 const execute = async (message, args, Course) => {
-  const courseName = args[0];
+  if (message.member.hasPermission("ADMINISTRATOR")) {
+    const courseName = args[0];
 
-  const guild = message.guild;
+    const guild = message.guild;
 
-  const courseString = findCategoryName(courseName, guild);
-  const category = guild.channels.cache.find(c => c.type === "category" && c.name === courseString);
+    const courseString = findCategoryName(courseName, guild);
+    const category = guild.channels.cache.find(c => c.type === "category" && c.name === courseString);
 
-  if (!category) return message.reply(`Error: Invalid course name: ${courseName}.`);
-  await Promise.all(guild.channels.cache
-    .filter(c => c.parent === category)
-    .map(async channel => await channel.delete()),
-  );
+    if (!category) return message.reply(`Error: Invalid course name: ${courseName}.`);
+    await Promise.all(guild.channels.cache
+      .filter(c => c.parent === category)
+      .map(async channel => await channel.delete()),
+    );
 
-  await category.delete();
+    await category.delete();
 
-  await Promise.all(guild.roles.cache
-    .filter(r => (r.name === `${courseName} ${courseAdminRole}` || r.name === courseName))
-    .map(async role => await role.delete()),
-  );
+    await Promise.all(guild.roles.cache
+      .filter(r => (r.name === `${courseName} ${courseAdminRole}` || r.name === courseName))
+      .map(async role => await role.delete()),
+    );
 
-  await removeCourseFromDb(courseName, Course);
-  await updateGuide(guild, Course);
+    await removeCourseFromDb(courseName, Course);
+    await updateGuide(guild, Course);
+  }
 };
 
 module.exports = {
